DAS-2347 - Move compute_source_swath back to resample.py since it is now only used there.
